@@ -1,24 +1,28 @@
 {
-  "name": "tika-text-extract",
+  "name": "@shelf/tika-text-extract",
   "version": "2.4.0",
-  "description": "Extract text from a document by Apache Tika",
+  "description": "Extract text from documents with Apache Tika",
+  "repository": "shelfio/tika-text-extract",
+  "license": "MIT",
+  "author": "Vlad Holubiev",
   "main": "lib/index.js",
   "files": [
     "lib/"
   ],
   "scripts": {
-    "build": "babel src --out-dir lib --ignore test.js",
+    "build": "rm -rf lib/ && babel src --out-dir lib --ignore 'src/*.test.ts','src/*.types.ts'  --extensions '.ts'",
     "coverage": "jest --coverage",
-    "lint": "eslint . --fix --format=node_modules/eslint-formatter-pretty",
+    "lint": "eslint . --ext .js,.ts,.json --fix",
     "prepublish": "npm run build",
-    "test": "jest "
+    "test": "jest"
   },
-  "engines": {
-    "node": ">=6"
+  "babel": {
+    "extends": "@shelf/babel-config/backend"
   },
-  "repository": "vladgolubev/tika-text-extract",
-  "author": "Vlad Holubiev",
-  "license": "MIT",
+  "prettier": "@shelf/prettier-config",
+  "jest": {
+    "testEnvironment": "node"
+  },
   "dependencies": {
     "debug": "^3.1.0",
     "get-stream": "^3.0.0",
@@ -27,16 +31,20 @@
     "is-stream": "^1.1.0"
   },
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-jest": "^22.0.0",
-    "babel-preset-env": "^1.6.1",
-    "eslint": "^4.13.1",
-    "eslint-config-google": "0.6.0",
-    "eslint-formatter-pretty": "^1.3.0",
-    "jest": "^22.0.0",
-    "memorystream": "^0.3.1"
+    "@babel/cli": "7.14.3",
+    "@babel/core": "7.14.3",
+    "@shelf/babel-config": "0.1.2",
+    "@shelf/eslint-config": "1.1.0",
+    "@shelf/prettier-config": "1.0.0",
+    "@shelf/tsconfig": "^0.0.6",
+    "@types/jest": "26.0.23",
+    "eslint": "7.27.0",
+    "jest": "27.0.1",
+    "memorystream": "^0.3.1",
+    "prettier": "2.3.0",
+    "typescript": "4.2.4"
   },
-  "jest": {
-    "testEnvironment": "node"
+  "engines": {
+    "node": ">=14"
   }
 }
