@@ -1,7 +1,8 @@
 const tte = require('../lib');
 const testFile = require('fs').readFileSync('./README.md');
 
-tte.startServer('/tmp/tika.jar')
+tte
+  .startServer('/tmp/tika.jar')
   .then(() => tte.extract(testFile))
   .then(console.log)
   .catch(console.log);
