@@ -7,13 +7,13 @@ const URL = 'http://localhost:9998/tika';
 
 /**
  * Extract text from a document
- * @param {Buffer|String|Promise} input File to extract text from
- * @return {Promise.<String>} Extracted text
+ * @param input File to extract text from
+ * @return Extracted text
  */
-export function extract(input = '') {
+export function extract(input: Buffer | string | Promise<string> = ''): Promise<string> {
   const fileStream = isStream(input) ? input : intoStream(input);
   const tikaStream = got.stream.put(URL, {
-    headers: {Accept: 'text/plain'}
+    headers: {Accept: 'text/plain'},
   });
 
   return getStream(fileStream.pipe(tikaStream));
